requirements.txt
openai==0.27.7
python-dotenv==1.0.0
mwclient==0.10.1
langchain==0.0.292
